Added authentication functionality for meals, recipes, and ingredients. TODO: Restructure Ingredients tables and rfactor server-side ingredients code

+++ capstone/java/src/main/java/com/techelevator/service/MealServiceImpl.java
@@ -6,10 +6,12 @@ import com.techelevator.exception.DaoException;
 import com.techelevator.exception.ServiceException;
 import com.techelevator.model.Meal;
 import com.techelevator.model.Mealplan;
+import com.techelevator.model.User;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.security.Principal;
 import java.util.List;
 
 @Component
@@ -32,17 +34,19 @@ public class MealServiceImpl implements MealService {
     }
 
 
-    public List<Meal> getMeals() throws InterruptedException {
+    public List<Meal> getMeals(Principal user) throws InterruptedException {
 //        Thread.sleep(2000); //Simulated loading time
-
-        return mealDao.listAllMeals();
+        User authUser = userDao.getUserByUsername(user.getName());
+        int userId = authUser.getId();
+        return mealDao.listAllMeals(userId);
     }
 
 
-    public Meal getMealById(int id) throws InterruptedException {
+    public Meal getMealById(int id, Principal user) throws InterruptedException {
 //        Thread.sleep(1000); //Simulated loading time
-
-        Meal result = mealDao.listMealById(id);
+        User authUser = userDao.getUserByUsername(user.getName());
+        int userId = authUser.getId();
+        Meal result = mealDao.listMealById(id, userId);
         if (result == null) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No meal with that id.");
         } else {
@@ -50,34 +54,53 @@ public class MealServiceImpl implements MealService {
         }
     }
 
-    public Meal createMeal(Meal meal) {
+    public Meal createMeal(Meal meal, Principal user) {
         try {
+            User authUser = userDao.getUserByUsername(user.getName());
+            meal.setUserId(authUser.getId());
             return mealDao.createMeal(meal);
         } catch (DaoException e) {
             throw new ServiceException("An error has occurred: " + e.getMessage());
         }
     }
 
-    public Meal updateMealInfo(int id, Meal updatedMeal) {
+    public Meal updateMealInfo(int id, Meal updatedMeal, Principal user) {
         updatedMeal.setMealId(id);
-        if (mealDao.updateMealInfo(updatedMeal)){
+        User authUser = userDao.getUserByUsername(user.getName());
+        int userId = authUser.getId();
+        if (mealDao.updateMealInfo(updatedMeal, userId)){
             return updatedMeal;
         } else {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Meal not found to update.");
         }
     }
 
-    public int addRecipeToMeal(int mealId, int recipeId) {
+    public int addRecipeToMeal(int mealId, int recipeId, Principal user) {
         try {
-            return mealDao.addRecipeToMeal(mealId, recipeId);
+            User authUser = userDao.getUserByUsername(user.getName());
+            int userId = authUser.getId();
+            Meal result = mealDao.listMealById(mealId, userId);
+            if(result != null){
+                return mealDao.addRecipeToMeal(mealId, recipeId);
+            }else{
+                throw new DaoException("Error: Cannot edit meal plan.");
+            }
         } catch (DaoException e) {
             throw new ServiceException("An error has occurred: " + e.getMessage());
         }
     }
 
-    public int removeRecipeFromMeal(int mealId, int recipeId) {
+    public int removeRecipeFromMeal(int mealId, int recipeId, Principal user) {
         try {
+            User authUser = userDao.getUserByUsername(user.getName());
+            int userId = authUser.getId();
+            Meal result = mealDao.listMealById(mealId, userId);
+            if(result !=null){
                 return mealDao.removeRecipeFromMeal(mealId, recipeId);
+
+            }else{
+                throw new DaoException("Error: Cannot edit meal plan.");
+            }
         } catch (DaoException e) {
                 throw new ServiceException("An error has occurred: " + e.getMessage());
         }
